getting_data/requirements.txt
paho-mqtt==1.6.1
psycopg2==2.9.7
python-decouple==3.8
